requirements.txt
certifi==2019.6.16
chardet==3.0.4
docker==4.0.2
idna==2.8
ovh==0.5.0
python-dateutil==2.8.0
raven==6.10.0
requests==2.22.0
six==1.12.0
urllib3==1.25.3
websocket-client==0.56.0
https://github.com/gijzelaerr/python-docker-machine/archive/v0.2.5.tar.gz#egg=python-docker-machine
ipython
pdbpp
